feat: 아이템 클릭 시 이벤트를 자식 요소에서 관리
