+++ packages/web/src/app/[lng]/(auth)/memos/_components/MemoView/MemoItem.tsx
@@ -2,10 +2,12 @@ import { GetMemoResponse } from '@extension/shared/types';
 import { cn } from '@extension/shared/utils';
 import { Card, CardContent } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import { HTMLAttributes, memo, useState } from 'react';
+import { HTMLAttributes, memo, MouseEvent, useState } from 'react';
 
 import MemoCardFooter from '../MemoCardFooter';
 import MemoCardHeader from '../MemoCardHeader';
+import { useRouter } from 'next/navigation';
+import { useSearchParams } from '@extension/shared/modules/search-params';
 
 interface MemoItemProps extends HTMLAttributes<HTMLElement>, LanguageType {
   memo: GetMemoResponse;
@@ -15,12 +17,26 @@ interface MemoItemProps extends HTMLAttributes<HTMLElement>, LanguageType {
 }
 
 export default memo(function MemoItem({ lng, memo, selectMemoItem, isSelecting, isSelected, ...props }: MemoItemProps) {
+  const router = useRouter();
+  const searchParams = useSearchParams();
   const [isHovered, setIsHovered] = useState(false);
+
   const handleMouseEnter = () => {
     setIsHovered(true);
   };
   const handleMouseLeave = () => {
     setIsHovered(false);
+  };
+
+  const handleMemoItemMouseDown = (event: MouseEvent<HTMLElement>) => {
+    event.stopPropagation();
+    const id = event.currentTarget.id;
+
+    if (isSelecting) selectMemoItem(Number(id));
+    else {
+      searchParams.set('id', id);
+      router.push(searchParams.getUrl(), { scroll: false });
+    }
   };
 
   return (
